pyproject.toml
[project]
name = "notion-task-mcp"
version = "0.1.0"
description = "Notion API를 활용한 Task 관리 MCP 서버"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "notion-client>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.scripts]
notion-task-mcp = "notion_task_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notion_task_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
# 외부 라이브러리 타입 스텁 이슈 무시
[[tool.mypy.overrides]]
module = ["notion_client.*", "mcp.*"]
ignore_missing_imports = true
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
